Remove CircuitOptimizer
