+++ test/CMakeLists.txt
@@ -7,7 +7,6 @@
 # Licensed under the MIT License
 
 # add various subdirectories containing tests
-add_subdirectory(circuit_optimizer)
 add_subdirectory(dd)
 add_subdirectory(ir)
 add_subdirectory(qdmi)
